refactor(graph): extract the source queries out of queries.ts (#1838)
The largest family, and the one that pulled `listAllSources` back where it
belonged: it sat among the tag queries but calls `collectSourceMetadata`, 400
lines further down. Moving it here is what makes both files honest.

Pure move again — six public functions plus nine private helpers, no signature
or logic changes, re-exported from `queries.ts`. Verified in both directions
before moving: the block calls nothing up-file, and nothing up-file calls into
it except that one `collectSourceMetadata` reference.

`DAY_MS` and `ReadingQueueView` moved with the reading-queue code that owns
them and are re-exported too, since `llm/approval` and `graph/index` import
them from here.
